+++ src/components/ProductCard/ProductCardFeatures.tsx
@@ -5,39 +5,29 @@ import './ProductCardFeatures.less';
 import ProductCardFeaturesTop from './ProductCardFeaturesTop';
 import ProductCardFeaturesBottom from './ProductCardFeaturesBottom';
 
-export interface ISocialIcon {
-    svgIcon: JSX.Element;
-    title: string;
-}
-
 export interface IShowcaseChildren {
-    svgIcon: JSX.Element;
     title: string;
-    caption: string;
-    value: string;
+    caption?: string;
+    svgIcon: JSX.Element;
 }
 
 interface IShowcaseParam {
     title?: string;
-    value?: string;
+    caption?: string;
     children: Array<Partial<IShowcaseChildren>>;
 }
 
 interface IProductCardFeaturesProps {
     /** Showcase first param
-     * Param with args: title: title: string(requred), value: string(requred), children: list of
-     * svgIcon: JSX.Element, title: string, caption: string, value: string
+     * Param with args: title: string(requred), caption: string(requred), children: list of
+     * svgIcon: JSX.Element, title: string
      */
     firstParam?: IShowcaseParam;
     /** Showcase second param
-     * Param with args: title: title: string(requred), value: string(requred), children: list of
-     * svgIcon: JSX.Element, title: string, caption: string, value: string
+     * Param with args: title: string, children: list of
+     * svgIcon: JSX.Element, title: string, caption: string
      */
     secondParam?: IShowcaseParam;
-    /** Social icons
-     * List with args: svgIcon: JSX.Element, title: string
-     */
-    socialIcons?: Array<Partial<ISocialIcon>>;
     className?: string;
     classNameTop?: string;
     classNameBottom?: string;
@@ -48,34 +38,24 @@ class ProductCardFeatures extends React.Component<IProductCardFeaturesProps, {}>
     static propTypes = {
         firstParam: PropTypes.shape({
             title: PropTypes.string,
-            value: PropTypes.string,
+            caption: PropTypes.string,
             children: PropTypes.arrayOf(
                 PropTypes.shape({
-                    svgIcon: PropTypes.element,
                     title: PropTypes.string,
-                    caption: PropTypes.string,
-                    value: PropTypes.string,
+                    svgIcon: PropTypes.element,
                 })
             ),
         }),
         secondParam: PropTypes.shape({
             title: PropTypes.string,
-            value: PropTypes.string,
             children: PropTypes.arrayOf(
                 PropTypes.shape({
-                    svgIcon: PropTypes.element,
                     title: PropTypes.string,
                     caption: PropTypes.string,
-                    value: PropTypes.string,
+                    svgIcon: PropTypes.element,
                 })
             ),
         }),
-        socialIcons: PropTypes.arrayOf(
-            PropTypes.shape({
-                svgIcon: PropTypes.element.isRequired,
-                title: PropTypes.string.isRequired,
-            })
-        ),
         className: PropTypes.string,
         classNameTop: PropTypes.string,
         classNameBottom: PropTypes.string,
@@ -89,15 +69,16 @@ class ProductCardFeatures extends React.Component<IProductCardFeaturesProps, {}>
                 {firstParam &&
                     <ProductCardFeaturesTop
                         className={cn('top', {}, classNameTop)}
+                        title={firstParam.title}
+                        caption={firstParam.caption}
                         params={firstParam.children}
-                        socialIcons={this.props.socialIcons}
                     />
                 }
                 {secondParam &&
                     <ProductCardFeaturesBottom
                         className={cn('bottom', {}, classNameBottom)}
+                        title={secondParam.title}
                         params={secondParam.children}
-                        title={secondParam.value}
                     />
                 }
             </div>